# smarthire-backend/specs/001-smarthire-backend-platform/contracts/openapi.yaml
openapi: "3.1.0"
info:
  title: SmartHire Backend API
  description: |
    Internal recruitment management platform API.
    All endpoints require a valid JWT bearer token except `/prescreen/**`, `/health`, `/docs`, and `/openapi.json`.
  version: "1.0.0"
  contact:
    name: SmartHire Platform Team
servers:
  - url: "http://localhost:8083"
    description: Local development
  - url: "https://smarthire.internal"
    description: Production
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ResponseDto:
      type: object
      properties:
        response:
          type: array
          items: {}
        message:
          type: string
        exception:
          type: string
          nullable: true
        result:
          type: string
          nullable: true
      required:
        - response
        - message
...
    RegisterUserRequest:
      type: object
      required: [employeeId, firstName, emailId, gradeId, buId, roleIds]
      properties:
        employeeId:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        emailId:
          type: string
          format: email
        gradeId:
          type: integer
          format: int64
        buId:
          type: integer
          format: int64
        practiceId:
          type: integer
          format: int64
        supervisorEmail:
          type: string
          format: email
        roleIds:
          type: array
          items:
            type: integer
            format: int64
        technologyIds:
          type: array
          items:
            type: integer
            format: int64
        accountIds:
          type: array
          items:
            type: integer
            format: int64
    UpdateRoleRequest:
      type: object
      required: [emailId, roleIds]
      properties:
        emailId:
          type: string
          format: email
        roleIds:
          type: array
          items:
            type: integer
            format: int64

    # -------------------------------------------------------------------------
    # Lookup
    # -------------------------------------------------------------------------
...
paths:
...
  /register/registerNewUser:
    post:
      tags: [UserManagement]
      summary: Register a new employee with roles and skills
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterUserRequest"
      responses:
        "200":
          description: User registered; welcome email sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseDto"
  /register/updateUserDetails:
    post:
      tags: [UserManagement]
      summary: Update an existing employee's details
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterUserRequest"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseDto"
  /register/removeSkill:
    post:
      tags: [UserManagement]
      summary: Remove a skill assignment from an employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [emailId, technologyId]
              properties:
                emailId:
                  type: string
                  format: email
                technologyId:
                  type: integer
                  format: int64
      responses:
        "200":
          description: Skill removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseDto"
  /role/getRole:
    post:
      tags: [UserManagement]
      summary: Get roles assigned to an employee by email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [emailId]
              properties:
                emailId:
                  type: string
                  format: email
      responses:
        "200":
          description: Employee roles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseDto"
  /users/getUsers:
    post:
      tags: [UserManagement]
      summary: List all users (optionally filtered)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                role:
                  type: string
                buId:
                  type: integer
                  format: int64
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseDto"
  /users/updateAssignedRole:
    post:
      tags: [UserManagement]
      summary: Update role assignments for an employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRoleRequest"
      responses:
        "200":
          description: Roles updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseDto"

  # ===========================================================================
  # Keycloak Integration
  # ===========================================================================
...